requirements.txt
# =====================================================
# CodeIt Ad SmallBiz - Backend Requirements
# 프로젝트 전체 의존성 통합 파일
# =====================================================

# ==================== FastAPI Backend ====================
fastapi==0.128.0
python-multipart==0.0.21  # 파일 업로드 지원
uvicorn[standard]==0.40.0

# ==================== Database ====================
alembic==1.12.0  # 마이그레이션
psycopg2-binary==2.9.11
sqlalchemy==2.0.45

# ==================== Data Validation ====================
email-validator==2.1.0
pydantic==2.12.5
pydantic-settings==2.12.0

# ==================== Authentication & Security ====================
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0

# ==================== Configuration ====================
python-dotenv==1.2.1
PyYAML==6.0.3

# ==================== OpenAI API ====================
openai==2.15.0

# ==================== RAG / LangChain ====================
chromadb==1.4.1
langchain==1.2.6
langchain-community==0.4.1
langchain-openai==1.1.7
sentence-transformers==5.2.0

# ==================== Core ML Libraries ====================
# 참고: PyTorch는 CUDA 버전에 따라 별도 설치 필요
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
# PyTorch: cu124 이상 필요
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0

accelerate==1.12.0
diffusers==0.36.0
transformers==4.57.3

# ==================== Model & Training ====================
peft==0.18.0
safetensors==0.7.0

# ==================== ControlNet ====================
controlnet-aux==0.0.10
mediapipe==0.10.9
timm==0.9.16

# ==================== Image Processing ====================
pillow==12.0.0
opencv-python==4.12.0.88
numpy==2.2.6
scipy==1.15.3
einops==0.8.1
scikit-image==0.25.2  # 이미지 품질 분석
rembg==2.0.69  # 배경 제거 (향후 전처리 노드)

# ==================== Utilities ====================
huggingface-hub==1.2.4
requests==2.32.5
tqdm==4.67.1

# ==================== Testing ====================
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx==0.28.1  # FastAPI 테스트용

# ==================== Development Tools ====================
black>=23.0.0  # 코드 포매팅
flake8>=6.0.0  # 린터
isort>=5.12.0  # import 정렬

# ==================== Optional (Production) ====================
# gunicorn>=21.2.0  # 프로덕션 WSGI 서버
# sentry-sdk>=1.40.0  # 에러 모니터링
